backend/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
flask_jwt_extended==4.7.1
flask_sqlalchemy==3.1.1
langchain==0.3.19
langchain_community==0.3.18
protobuf==5.29.3
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Werkzeug==3.1.3
